Breakout settings and prepare for dynamic paths
